refactor(me): dashboardの型と表示コンポーネントを分割
- me-dashboard.tsx から型定義を app/me/types.ts に移動
- ActivityHeatmap / AttemptDetailView / SummaryCard を分離
- 日付フォーマット関数を app/me/utils.ts に分離
- 既存UI仕様を維持しつつ責務を分離

closes #109

Made-with: Cursor
