+++ app/me/me-dashboard.tsx
@@ -3,102 +3,25 @@
 import { useRouter } from "next/navigation";
 import { useEffect, useRef, useState } from "react";
 
-type CategoryScore = {
-  category: string;
-  total: number;
-  correct: number;
-  percent: number;
-};
-
-type AttemptFilters = {
-  categories?: string[];
-  level?: number;
-  count?: number;
-};
-
-type AttemptSummary = {
-  id: string;
-  status: "IN_PROGRESS" | "COMPLETED";
-  filters: AttemptFilters;
-  startedAt: string;
-  completedAt: string | null;
-  result: {
-    overallPercent: number;
-    categoryBreakdown: CategoryScore[];
-  } | null;
-};
-
-type QuestionDetail = {
-  attemptQuestionId: string;
-  order: number;
-  selectedIndex: number | null;
-  isCorrect: boolean | null;
-  question: {
-    id: string;
-    category: string;
-    level: number;
-    questionText: string;
-    choices: string[];
-    answerIndex?: number;
-    explanation?: string;
-  };
-};
-
-type AttemptDetail = AttemptSummary & {
-  questions: QuestionDetail[];
-};
-
-type NotionDeliveryState = {
-  isSending: boolean;
-  message: string;
-  kind: "success" | "error" | null;
-};
-
-type ExportState = {
-  isExporting: boolean;
-  message: string;
-  kind: "success" | "error" | null;
-};
-
-type MeStats = {
-  totalAttempts: number;
-  averagePercent: number;
-  recentAveragePercent: number;
-  bestPercent: number;
-  streakDays: number;
-  totalAnswered: number;
-  recent7DaysAnswered: number;
-  weeklyActivity: Array<{
-    date: string;
-    count: number;
-  }>;
-  activityHeatmap: Array<{
-    date: string;
-    count: number;
-  }>;
-  categoryProgress: CategoryScore[];
-};
-
-type MeProfile = {
-  id: string;
-  email: string;
-};
-
-type MeTabKey = "summary" | "history";
+import { ActivityHeatmap } from "./components/activity-heatmap";
+import { AttemptDetailView } from "./components/attempt-detail-view";
+import { SummaryCard } from "./components/summary-card";
+import type {
+  AttemptDetail,
+  AttemptFilters,
+  AttemptSummary,
+  ExportState,
+  MeProfile,
+  MeStats,
+  MeTabKey,
+  NotionDeliveryJobSnapshot,
+  NotionDeliveryState,
+} from "./types";
+import { formatDate } from "./utils";
 
 const ATTEMPTS_PAGE_SIZE = 10;
 const NOTION_STATUS_POLL_INTERVAL_MS = 1500;
 const NOTION_STATUS_POLL_MAX_ATTEMPTS = 40;
-
-type NotionDeliveryJobSnapshot = {
-  id: string;
-  totalQuestions: number;
-  processedQuestions: number;
-  successQuestions: number;
-  failedQuestions: number;
-  duplicateDetected: boolean;
-  lastError: string | null;
-};
 
 export const MeDashboard = () => {
   const router = useRouter();
@@ -975,315 +898,3 @@ export const MeDashboard = () => {
   );
 };
 
-const ActivityHeatmap = ({
-  items,
-}: {
-  items: Array<{ date: string; count: number }>;
-}) => {
-  const toDateKey = (date: Date): string => date.toISOString().slice(0, 10);
-  const activityByDate = new Map(items.map((item) => [item.date, item.count]));
-  const totalContributions = items.reduce((sum, item) => sum + item.count, 0);
-  const today = new Date();
-  today.setUTCHours(0, 0, 0, 0);
-  const oneYearAgo = new Date(today);
-  oneYearAgo.setUTCDate(today.getUTCDate() - 364);
-  const firstGridDate = new Date(oneYearAgo);
-  const mondayOffset = (firstGridDate.getUTCDay() + 6) % 7;
-  firstGridDate.setUTCDate(firstGridDate.getUTCDate() - mondayOffset);
-
-  const columns: Array<
-    Array<{ date: string; count: number; isInRange: boolean }>
-  > = [];
-  const cursor = new Date(firstGridDate);
-  while (cursor <= today) {
-    const week: Array<{ date: string; count: number; isInRange: boolean }> = [];
-    for (let dayIndex = 0; dayIndex < 7; dayIndex += 1) {
-      const key = toDateKey(cursor);
-      const isInRange = cursor >= oneYearAgo && cursor <= today;
-      week.push({
-        date: key,
-        count: isInRange ? (activityByDate.get(key) ?? 0) : 0,
-        isInRange,
-      });
-      cursor.setUTCDate(cursor.getUTCDate() + 1);
-    }
-    columns.push(week);
-  }
-
-  const weekCount = columns.length;
-  const heatmapCellSize = 10;
-  const heatmapGapSize = 2;
-  const heatmapGridTemplate = `repeat(${weekCount}, ${heatmapCellSize}px)`;
-  const maxCount = Math.max(
-    ...columns.flatMap((week) =>
-      week.filter((day) => day.isInRange).map((day) => day.count),
-    ),
-    1,
-  );
-
-  const monthLabels = columns.map((week, weekIndex) => {
-    const firstInRangeDay = week.find((day) => day.isInRange);
-    if (!firstInRangeDay) {
-      return "";
-    }
-    const date = new Date(firstInRangeDay.date);
-    const month = date.getUTCMonth();
-    const prevFirstDay = columns[weekIndex - 1]?.find((day) => day.isInRange);
-    const prevMonth =
-      prevFirstDay === undefined
-        ? -1
-        : new Date(prevFirstDay.date).getUTCMonth();
-    return month !== prevMonth
-      ? date.toLocaleDateString("en-US", { month: "short", timeZone: "UTC" })
-      : "";
-  });
-  const weekdayLabels: string[] = ["Mon", "", "Wed", "", "Fri", "", ""];
-
-  return (
-    <article className="rounded-2xl border border-black/10 bg-white p-6 dark:border-white/15 dark:bg-black/50">
-      <h2 className="mb-1 text-lg font-semibold">学習アクティビティ</h2>
-      <p className="mb-4 text-sm text-neutral-600 dark:text-neutral-400">
-        {totalContributions.toLocaleString("en-US")} contributions in the last year
-      </p>
-      <div className="overflow-x-auto pb-1">
-        <div className="min-w-max rounded-lg border border-neutral-200 bg-neutral-50 p-2 dark:border-neutral-700 dark:bg-neutral-900/40">
-          <div className="flex flex-col gap-2">
-            <div
-              className="ml-7 grid"
-              style={{
-                gridTemplateColumns: heatmapGridTemplate,
-                columnGap: `${heatmapGapSize}px`,
-              }}
-            >
-              {monthLabels.map((label, index) => (
-                <span
-                  key={`${label}-${index}`}
-                  className="text-[10px] leading-none text-neutral-500 dark:text-neutral-400"
-                >
-                  {label}
-                </span>
-              ))}
-            </div>
-            <div className="flex gap-1.5">
-              <div className="grid grid-rows-7 items-center text-[10px] text-neutral-500 dark:text-neutral-400">
-                {weekdayLabels.map((label, index) => (
-                  <span key={`${label}-${index}`} className="h-2.5 leading-none">
-                    {label}
-                  </span>
-                ))}
-              </div>
-              <div className="inline-flex gap-0.5">
-                {columns.map((week, columnIndex) => (
-                  <div key={columnIndex} className="grid grid-rows-7 gap-0.5">
-                    {week.map((item) => {
-                      if (!item.isInRange) {
-                        return (
-                          <div
-                            key={item.date}
-                            className="h-2.5 w-2.5 rounded-[2px] bg-transparent"
-                            aria-hidden="true"
-                          />
-                        );
-                      }
-
-                      const level = getHeatLevel(item.count, maxCount);
-                      return (
-                        <div
-                          key={item.date}
-                          title={`${item.date}: ${item.count}問`}
-                          aria-label={`${item.date}に${item.count}問回答`}
-                          className={`h-2.5 w-2.5 rounded-[2px] ${level}`}
-                        />
-                      );
-                    })}
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="mt-3 flex items-center justify-end gap-2 text-xs text-neutral-500 dark:text-neutral-400">
-        <span>Less</span>
-        <div className="flex items-center gap-1">
-          <span className="h-3 w-3 rounded-sm bg-neutral-200 dark:bg-neutral-700" />
-          <span className="h-3 w-3 rounded-sm bg-brand-200 dark:bg-brand-400/35" />
-          <span className="h-3 w-3 rounded-sm bg-brand-300 dark:bg-brand-400/60" />
-          <span className="h-3 w-3 rounded-sm bg-brand-500 dark:bg-brand-300" />
-        </div>
-        <span>More</span>
-      </div>
-    </article>
-  );
-};
-
-const AttemptDetailView = ({ attempt }: { attempt: AttemptDetail }) => {
-  const incorrectQuestions = attempt.questions.filter(
-    (q) => q.isCorrect === false,
-  );
-
-  return (
-    <section className="space-y-4">
-      <h2 className="text-lg font-semibold">
-        テスト詳細
-        {attempt.result && (
-          <span className="ml-2 text-brand-600 dark:text-brand-300">
-            {attempt.result.overallPercent}%
-          </span>
-        )}
-      </h2>
-
-      {/* カテゴリ別スコア */}
-      {attempt.result && (
-        <div className="rounded-2xl border border-black/10 bg-white p-6 dark:border-white/15 dark:bg-black/50">
-          <h3 className="mb-3 text-sm font-medium text-neutral-700 dark:text-neutral-300">
-            カテゴリ別正答率
-          </h3>
-          <div className="space-y-2">
-            {(attempt.result.categoryBreakdown as CategoryScore[]).map((cat) => (
-              <div key={cat.category} className="flex items-center gap-3">
-                <span className="w-28 shrink-0 text-sm text-neutral-600 dark:text-neutral-400">
-                  {cat.category}
-                </span>
-                <div className="flex-1">
-                  <div className="h-2.5 overflow-hidden rounded-full bg-neutral-200 dark:bg-neutral-700">
-                    <div
-                      className="h-full rounded-full bg-brand-400 transition-all dark:bg-brand-300"
-                      style={{ width: `${cat.percent}%` }}
-                    />
-                  </div>
-                </div>
-                <span className="w-20 text-right text-sm font-medium">
-                  {cat.correct}/{cat.total} ({cat.percent}%)
-                </span>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-
-      {/* 間違い問題一覧 */}
-      {incorrectQuestions.length > 0 && (
-        <div className="space-y-3">
-          <h3 className="text-sm font-medium text-neutral-700 dark:text-neutral-300">
-            間違えた問題 ({incorrectQuestions.length}件)
-          </h3>
-          {incorrectQuestions.map((q) => (
-            <div
-              key={q.attemptQuestionId}
-              className="rounded-2xl border border-black/10 bg-white p-5 dark:border-white/15 dark:bg-black/50"
-            >
-              <div className="mb-2 flex items-center gap-2 text-xs text-neutral-500 dark:text-neutral-400">
-                <span className="rounded bg-neutral-100 px-2 py-0.5 dark:bg-neutral-800">
-                  Q{q.order}
-                </span>
-                <span className="rounded bg-neutral-100 px-2 py-0.5 dark:bg-neutral-800">
-                  {q.question.category}
-                </span>
-                <span className="rounded bg-red-100 px-2 py-0.5 text-red-700 dark:bg-red-900/30 dark:text-red-400">
-                  不正解
-                </span>
-              </div>
-
-              <p className="mb-3 text-sm leading-relaxed">{q.question.questionText}</p>
-
-              <div className="mb-3 flex flex-col gap-1.5">
-                {(q.question.choices as string[]).map((choice, i) => {
-                  const isAnswer = i === q.question.answerIndex;
-                  const isUserChoice = i === q.selectedIndex;
-
-                  return (
-                    <div
-                      key={i}
-                      className={`rounded-lg px-3 py-2 text-sm ${
-                        isAnswer
-                          ? "border border-green-300 bg-green-50 dark:border-green-700 dark:bg-green-900/20"
-                          : isUserChoice
-                            ? "border border-red-300 bg-red-50 dark:border-red-700 dark:bg-red-900/20"
-                            : "border border-transparent"
-                      }`}
-                    >
-                      <span className="mr-2 font-medium">
-                        {String.fromCharCode(65 + i)}.
-                      </span>
-                      {choice}
-                      {isAnswer && (
-                        <span className="ml-2 text-xs text-green-600 dark:text-green-400">
-                          正解
-                        </span>
-                      )}
-                      {isUserChoice && !isAnswer && (
-                        <span className="ml-2 text-xs text-red-600 dark:text-red-400">
-                          あなたの回答
-                        </span>
-                      )}
-                    </div>
-                  );
-                })}
-              </div>
-
-              {q.question.explanation && (
-                <div className="rounded-lg bg-neutral-50 px-3 py-2 text-sm text-neutral-600 dark:bg-neutral-800/50 dark:text-neutral-400">
-                  <span className="font-medium">解説: </span>
-                  {q.question.explanation}
-                </div>
-              )}
-            </div>
-          ))}
-        </div>
-      )}
-
-      {incorrectQuestions.length === 0 && attempt.status === "COMPLETED" && (
-        <div className="rounded-2xl border border-black/10 bg-white p-6 dark:border-white/15 dark:bg-black/50">
-          <p className="text-sm text-green-600 dark:text-green-400">
-            全問正解です！
-          </p>
-        </div>
-      )}
-    </section>
-  );
-};
-
-const SummaryCard = ({
-  label,
-  value,
-}: {
-  label: string;
-  value: string;
-}) => {
-  return (
-    <article className="rounded-xl border border-black/10 bg-white p-4 dark:border-white/15 dark:bg-black/50">
-      <p className="text-xs text-neutral-500 dark:text-neutral-400">{label}</p>
-      <p className="mt-2 text-2xl font-semibold text-neutral-900 dark:text-neutral-100">
-        {value}
-      </p>
-    </article>
-  );
-};
-
-const getHeatLevel = (count: number, maxCount: number): string => {
-  if (count === 0) {
-    return "bg-neutral-200 dark:bg-neutral-700";
-  }
-  if (count <= maxCount * 0.33) {
-    return "bg-brand-200 dark:bg-brand-400/35";
-  }
-  if (count <= maxCount * 0.66) {
-    return "bg-brand-300 dark:bg-brand-400/60";
-  }
-  return "bg-brand-500 dark:bg-brand-300";
-};
-
-const formatDate = (dateString: string | null): string => {
-  if (!dateString) return "";
-
-  const date = new Date(dateString);
-
-  return date.toLocaleDateString("ja-JP", {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-    hour: "2-digit",
-    minute: "2-digit",
-  });
-};
